Fixed unable to set PWM duty cycle to 100% (256)

+++ Firmware/VibHub-ESP32/Pwm.h
@@ -14,7 +14,7 @@ class Pwm{
         Pwm(uint8_t pin);
         Pwm();  // Don't initialize immediately
         void initialize(uint8_t pin);
-        void setPWM(uint8_t duty);
+        void setPWM(uint16_t duty);
         static void disableAll();
         int _channel;
